Changed the sound backend interface to allow safe vectorization with a known padded buffer size.
